requirements.txt
google-genai>=1.0.0
openai>=1.0.0
pyyaml>=6.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
